+++ Helpers/Models/CasperServiceModels/JobOfferDetailed.cs
@@ -10,7 +10,7 @@ namespace Helpers.Models.CasperServiceModels
     {
         public int auction_type_id { get; set; }
         public string deploy_hash { get; set; }
-        public int expected_time_frame { get; set; }
+        public long expected_time_frame { get; set; }
         public int job_offer_id { get; set; }
         public string job_poster { get; set; }
         public long max_budget { get; set; }
